move `ui.context_menu` into separate file
