+++ nicegui/ui.py
@@ -18,6 +18,7 @@ __all__ = [
     'color_picker',
     'colors',
     'column',
+    'context_menu',
     'dark_mode',
     'date',
     'dialog',
@@ -40,7 +41,6 @@ __all__ = [
     'link_target',
     'log',
     'markdown',
-    'context_menu',
     'menu',
     'menu_item',
     'mermaid',
@@ -116,6 +116,7 @@ from .elements.color_input import ColorInput as color_input
 from .elements.color_picker import ColorPicker as color_picker
 from .elements.colors import Colors as colors
 from .elements.column import Column as column
+from .elements.context_menu import ContextMenu as context_menu
 from .elements.dark_mode import DarkMode as dark_mode
 from .elements.date import Date as date
 from .elements.dialog import Dialog as dialog
@@ -138,7 +139,6 @@ from .elements.link import Link as link
 from .elements.link import LinkTarget as link_target
 from .elements.log import Log as log
 from .elements.markdown import Markdown as markdown
-from .elements.menu import ContextMenu as context_menu
 from .elements.menu import Menu as menu
 from .elements.menu import MenuItem as menu_item
 from .elements.mermaid import Mermaid as mermaid
